refactor: updating the project structure for bootstrap
