+++ apps/ubuntu_bootstrap/lib/app.dart
@@ -11,9 +11,9 @@ import 'package:path/path.dart' as p;
 import 'package:subiquity_client/subiquity_client.dart';
 import 'package:subiquity_client/subiquity_server.dart';
 import 'package:timezone_map/timezone_map.dart';
-import 'package:ubuntu_bootstrap/installer/installation_step.dart';
-import 'package:ubuntu_bootstrap/installer/installer_model.dart';
-import 'package:ubuntu_bootstrap/installer/installer_wizard.dart';
+import 'package:ubuntu_bootstrap/app/installation_step.dart';
+import 'package:ubuntu_bootstrap/app/installer_model.dart';
+import 'package:ubuntu_bootstrap/app/installer_wizard.dart';
 import 'package:ubuntu_bootstrap/l10n.dart';
 import 'package:ubuntu_bootstrap/pages/loading/loading_page.dart';
 import 'package:ubuntu_bootstrap/services.dart';
@@ -24,7 +24,7 @@ import 'package:ubuntu_utils/ubuntu_utils.dart';
 import 'package:ubuntu_wizard/ubuntu_wizard.dart';
 import 'package:yaru/yaru.dart';
 
-export 'installer/installer_wizard.dart';
+export 'app/installer_wizard.dart';
 
 Future<void> runInstallerApp(
   List<String> args, {
