+++ apps/ubuntu_bootstrap/lib/app/installer_model.dart
