+++ apps/ubuntu_bootstrap/lib/main.dart
@@ -1,3 +1,3 @@
-import 'package:ubuntu_bootstrap/installer.dart';
+import 'package:ubuntu_bootstrap/app.dart';
 
 Future<void> main(List<String> args) => runInstallerApp(args);
